Rename `image_surface_png.rs` => `surface_png.rs`
